+++ cypress/tests/core/blocks-image.js
@@ -106,7 +106,7 @@ describe('Blocks Tests', () => {
     // then image src must be equal to image name
     cy.get('.block img')
       .should('have.attr', 'src')
-      .and('eq', '/my-page/image.png/@@images/image');
+      .and('eq', '/my-page/image.png/@@images/image/listing');
   });
 
   it('Create a image block document in edit mode', () => {