+++ src/config/index.js
@@ -31,6 +31,7 @@ import { loadables } from './Loadables';
 
 import { sentryOptions } from './Sentry';
 import { contentIcons } from './ContentIcons';
+import { imageScales } from './ImageScales';
 
 import applyAddonConfiguration from 'load-volto-addons';
 
@@ -116,7 +117,8 @@ let config = {
     sentryOptions: {
       ...sentryOptions,
     },
-    contentIcons: contentIcons,
+    contentIcons,
+    imageScales,
     loadables,
     lazyBundles: {
       cms: [