+++ src/helpers/index.js
@@ -69,6 +69,8 @@ export {
   applyConfig,
   withServerErrorCode,
 } from '@plone/volto/helpers/Utils/Utils';
+
+export { getImageAttributes } from './Image/Image';
 export { messages } from './MessageLabels/MessageLabels';
 export { asyncConnect } from './AsyncConnect';
 export {